v4l: Add support for V4L2_BUF_TYPE_META_OUTPUT
The V4L2_BUF_TYPE_META_OUTPUT mirrors the V4L2_BUF_TYPE_META_CAPTURE
with the exception that it is an OUTPUT type. The use case for this
is to pass buffers to the device that are not image data but metadata.
The formats, just as the metadata capture formats, are typically
device specific and highly structured.

Change-Id: Ib7f5075f22737290ddc3dcd53de737a7be2a09e4
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tian Shu Qiu <tian.shu.qiu@intel.com>